backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pymongo==4.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
scikit-learn==1.3.2
joblib==1.3.2
google-cloud-dialogflow==2.17.0
google-cloud-dialogflow-cx==1.29.0
requests==2.31.0
aiofiles==23.2.1
cors==1.0.1